# Bai1/github-copilot-vibe-coding-workshop/openapi.yaml
openapi: 3.0.1
info:
  title: Simple Social Media API
  description: A basic Social Networking Service (SNS) API that allows users to create, retrieve, update, and delete posts; add comments; and like/unlike posts.
  version: 1.0.0
  contact:
    name: Contoso Product Team
    email: support@contoso.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api
    description: Development server
paths:
  /posts:
    get:
      summary: List all posts
      description: Retrieve a list of all posts in reverse chronological order for browsing.
      operationId: listPosts
      tags:
        - Posts
      responses:
        '200':
          description: Successful response with list of posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Post'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Create a new post
      description: Create a new post to share content with others.
      operationId: createPost
      tags:
        - Posts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePostRequest'
      responses:
        '201':
          description: Post created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          description: Bad request - invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /posts/{postId}:
    get:
      summary: Get a single post
      description: Retrieve details of a specific post by its ID.
      operationId: getPost
      tags:
        - Posts
      parameters:
        - name: postId
          in: path
          required: true
          description: The ID of the post to retrieve
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful response with post details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '404':
          description: Post not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      summary: Update a post
      description: Update the content of an existing post.
      operationId: updatePost
      tags:
        - Posts
      parameters:
        - name: postId
          in: path
          required: true
          description: The ID of the post to update
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePostRequest'
      responses:
        '200':
          description: Post updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          description: Bad request - invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Post not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete a post
      description: Delete a specific post by its ID.
      operationId: deletePost
      tags:
        - Posts
      parameters:
        - name: postId
          in: path
          required: true
          description: The ID of the post to delete
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Post deleted successfully
        '404':
          description: Post not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /posts/{postId}/comments:
    get:
      summary: List comments for a post
      description: Retrieve all comments for a specific post.
      operationId: listComments
      tags:
        - Comments
      parameters:
        - name: postId
          in: path
          required: true
          description: The ID of the post
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful response with list of comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Comment'
        '404':
          description: Post not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Create a comment on a post
      description: Add a new comment to a specific post to share thoughts.
      operationId: createComment
      tags:
        - Comments
      parameters:
        - name: postId
          in: path
          required: true
          description: The ID of the post
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommentRequest'
      responses:
        '201':
          description: Comment created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '400':
          description: Bad request - invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Post not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /posts/{postId}/comments/{commentId}:
    get:
      summary: Get a specific comment
      description: Retrieve details of a specific comment by its ID.
      operationId: getComment
      tags:
        - Comments
      parameters:
        - name: postId
          in: path
          required: true
          description: The ID of the post
          schema:
            type: integer
            format: int64
        - name: commentId
          in: path
          required: true
          description: The ID of the comment
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful response with comment details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '404':
          description: Comment or post not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      summary: Update a comment
      description: Update the content of an existing comment.
      operationId: updateComment
      tags:
        - Comments
      parameters:
        - name: postId
          in: path
          required: true
          description: The ID of the post
          schema:
            type: integer
            format: int64
        - name: commentId
          in: path
          required: true
          description: The ID of the comment
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCommentRequest'
      responses:
        '200':
          description: Comment updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '400':
          description: Bad request - invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Comment or post not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete a comment
      description: Delete a specific comment by its ID.
      operationId: deleteComment
      tags:
        - Comments
      parameters:
        - name: postId
          in: path
          required: true
          description: The ID of the post
          schema:
            type: integer
            format: int64
        - name: commentId
          in: path
          required: true
          description: The ID of the comment
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Comment deleted successfully
        '404':
          description: Comment or post not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /posts/{postId}/likes:
    post:
      summary: Like a post
      description: Add a like to a specific post to show appreciation.
      operationId: likePost
      tags:
        - Likes
      parameters:
        - name: postId
          in: path
          required: true
          description: The ID of the post to like
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LikeRequest'
      responses:
        '201':
          description: Like added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Like'
        '400':
          description: Bad request - invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Post not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Unlike a post
      description: Remove a like from a specific post.
      operationId: unlikePost
      tags:
        - Likes
      parameters:
        - name: postId
          in: path
          required: true
          description: The ID of the post to unlike
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LikeRequest'
      responses:
        '204':
          description: Like removed successfully
        '404':
          description: Post or like not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
    Post:
      type: object
      description: A social media post
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier for the post
          example: 1
        username:
          type: string
          description: Username of the post author
          example: "john_doe"
        content:
          type: string
          description: Content of the post
          example: "Just had an amazing hiking experience in the mountains!"
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the post was created
          example: "2025-05-30T12:00:00Z"
        updatedAt:
          type: string
          format: date-time
          description: Timestamp when the post was last updated
          example: "2025-05-30T12:00:00Z"
        likesCount:
          type: integer
          description: Number of likes on the post
          example: 5
        commentsCount:
          type: integer
          description: Number of comments on the post
          example: 3
      required:
        - id
        - username
        - content
        - createdAt
        - updatedAt
        - likesCount
        - commentsCount
    CreatePostRequest:
      type: object
      description: Request payload for creating a new post
      properties:
        username:
          type: string
          description: Username of the post author
          example: "john_doe"
        content:
          type: string
          description: Content of the post
          example: "Just had an amazing hiking experience in the mountains!"
      required:
        - username
        - content
    UpdatePostRequest:
      type: object
      description: Request payload for updating an existing post
      properties:
        username:
          type: string
          description: Username of the post author
          example: "john_doe"
        content:
          type: string
          description: Updated content of the post
          example: "Updated: Just had an amazing hiking experience in the mountains with friends!"
      required:
        - username
        - content
    Comment:
      type: object
      description: A comment on a post
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier for the comment
          example: 1
        postId:
          type: integer
          format: int64
          description: ID of the post this comment belongs to
          example: 1
        username:
          type: string
          description: Username of the comment author
          example: "jane_smith"
        content:
          type: string
          description: Content of the comment
          example: "That sounds incredible! Which trail did you take?"
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the comment was created
          example: "2025-05-30T13:00:00Z"
        updatedAt:
          type: string
          format: date-time
          description: Timestamp when the comment was last updated
          example: "2025-05-30T13:00:00Z"
      required:
        - id
        - postId
        - username
        - content
        - createdAt
        - updatedAt
    CreateCommentRequest:
      type: object
      description: Request payload for creating a new comment
      properties:
        username:
          type: string
          description: Username of the comment author
          example: "jane_smith"
        content:
          type: string
          description: Content of the comment
          example: "That sounds incredible! Which trail did you take?"
      required:
        - username
        - content
    UpdateCommentRequest:
      type: object
      description: Request payload for updating an existing comment
      properties:
        username:
          type: string
          description: Username of the comment author
          example: "jane_smith"
        content:
          type: string
          description: Updated content of the comment
          example: "That sounds incredible! Which trail did you take? I'd love to try it myself."
      required:
        - username
        - content
...
    Error:
      type: object
      description: Error response
      properties:
        message:
          type: string
          description: Error message describing what went wrong
          example: "Post not found"
      required:
        - message